requirements.txt
openai>=1.40.0
opencv-python>=4.9.0
Pillow>=10.0.0
PyYAML>=6.0.1
jsonschema>=4.23.0
pytest>=8.0.0
python-dotenv>=1.0.1
